voice-agent/pyproject.toml
[project]
name = "agent-matcha-voice"
version = "0.1.0"
description = "Club Allenby voice intake agent powered by LiveKit"
requires-python = ">=3.11"
dependencies = [
    "livekit-agents>=1.0.0",
    "livekit-plugins-deepgram>=1.0.0",
    # "livekit-plugins-cartesia>=1.0.0",  # TODO: re-enable when switching from Deepgram TTS
    "livekit-plugins-openai>=1.0.0",
    "livekit-plugins-silero>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
agent = "agent:main"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
